refactor(dapp): Use enums for network symbols
Without this patch, network symbols will be used as `strings`.

This is a problem because it might cause a problem if there is a typo or if we
have a lot of network symbols as is expected when the project grows.

This patch solves the problem by using `enums` for network symbols and use it
for network logos as well.

Issue-URL: https://app.clickup.com/t/86699tqan