fix(dapp): clear staking input when changing account
Without this patch, the value inside the staking input is not cleared
when we change the wallet account from metamask.

This is a problem because the value that is already in the input field
may be greater than what the new chosen wallet account have.

This patch solves the problem by resetting the stakeAmount value whenever
the wallet account changed.

Issue-URL: https://app.clickup.com/t/861m4jtgc