fix(dapp): mobile navbar switch on when wallet disconnected
Without this patch, when wallet is disconnected
our mobile navbar design will be shown when
we still have free space.

This is a problem because we must show mobile
navbar at a smalled screen size.

This patch solves the problem by adding some
conditional classed on the navbar elements
in order to fix the issue.

Issue-URL: https://app.clickup.com/t/861m7gppx