feat(dapp): show testing background
Without this patch, the dapp will not show that a given build is only used for
testing purposes

This is a problem because some of our dev builds will be deployed on IPFS. A
regular user might attempt to use it but with the testing background we can
alert the user.

This patch solves the problem by adding a testing background image in the
staging environment.

Issue-URL: https://app.clickup.com/t/8669c84m2