feat(root): add dangerjs automation
Without this patch, there is no way to check the commit messages and the
quality of the submitted merge requests in an automated way.

This is a problem because it takes time for the reviewer and is
unproductive for the whole team.

This patch solves the problem by adding dangerjs to the project.
Dangerjs runs during the CI process and automates everyday code review
chores. The dangerjs rules include the following checks:
- the author does not follow conventional commits specification
- simultaneous change of the package.json and yarn.lock
- lines in commits are longer than 80 characters
- commits have an Issue-URL
- docs are updated
- tests are added
- MR is quite large
- the presence of MR description

Issue-URL: https://app.clickup.com/t/296rg3u