feat(dapp): keep previous builds on IPFS for every MR
Without this patch, with every push to the remote branch, we will re-build the
dapp and re-deploy it to IPFS.

This is a problem because we will lose the previous build with every push.

This patch solves the problem by storing all previous builds in the GitLab
comment in JSON string format. Using this approach we will be able to display
all previous build's commits SHA and its corresponding IPFS URL.

Issue-URL: https://app.clickup.com/t/8669db8yh