refactor(dapp): update the data model of the dApp
Without this patch, the dApp data model focused on advanced staking. To
facilitate the development of dApp v.1.0 and of the DeFi adaptor, we
need to update the data model of the dApp and prepare it for the
transition from the advanced staking dApp v.0.5 to Panther Wallet v.1.0.

This patch solves the problem by updating the data model of the dApp.
More specifically, it renames the AdvancedStakesRewards slice in the
redux to the utxos, removes PRPs from the slice, adds new types, and
renames function names. It also updates the components to use the new
UTXOs slice and types.

Issue-URL: https://app.clickup.com/t/860prhkj8