feat(dapp): use ipfs script to deploy to any ipfs node
Without this patch, we can only use the IPFS deployment script to deploy to
Infura node and from the CI/CD only. Also, we will have a duplicated logic with
the original bash script

This is a problem because we should reduce the logic duplication and use the
same script for multiple purposes.

This patch solves the problem by updating the IPFS script to deploy to any IPFS
node (local, remote, Infura) using basic node info like host, port, protocol.

Issue-URL: https://app.clickup.com/t/8669fh0d5