docs(dapp): add docs for the ipfs script
Without this patch, we will not have any document that explains how to use the
`deploy-ipfs.ts` script

This is a problem because the script has multiple use cases and can be used in
different setting and with every use case it expects a different set of
enviornemental variables.

This patch solves the problem by adding a markdown file that describes it in
detail the use cases for the scripts and what parameters the script expects.

Issue-URL: https://app.clickup.com/t/8669fh0d5