fix(dapp): the PRP rewards calculation after 2000 stakes
Without this patch, the PRP rewards calculation is incorrect after 2000
stakes. It should show 0 rewards for all stakes after 2000.

This patch solves the problem by updating a function to calculate the
PRP rewards for stakes after 2000. It uses the timestamp of the last
reward generated on the chain: Polygon txHash:
0x3314b6a4c89e300c55dcee276699fc92dae64a6882406bc836a138467201c10f. The
logic that relied on the leaf number was removed. This patch also
updates the tests.

Issue-URL: https://app.clickup.com/t/861mba0kn