deploy(root): remove hardcoded addresses in adv. staking setup script
Without this patch, the setup script contains the hardcoded address of
the deployed PNftToken contract.

This is a problem because, on the new testing network deployments, the
address of the smart contract changes. As a result, the script may fail
to configure the smart contracts.

This patch solves the problem by replacing the hardcoded address of the
smart contracts with the placeholder in the setup.sh script.
Additionally, it updates the comments with helpful info.

Issue-URL: https://app.clickup.com/t/861mbq4zn