test(dapp): setup synpress [e2e testing framework]
Without this patch, we will not have any e2e testing to test the dapp

This is a problem because we need to ensure that all the dapp pieces/features
work together. Also any new change might break existing features.

This patch solves the problem by using Synpress
https://github.com/Synthetixio/synpress. Synpress is e2e testing framework based
on Cypress and playwright with support for metamask.

Issue-URL: https://app.clickup.com/t/8669jq330