docs(circuits): addition of readme file to circuit folder
Without this patch there won't be any documentation on how to install,
set-up and run the project.

This is a problem because without documentation it is hard to review and
understand the code.

This patch solves the problem by providing documentation on how to set-
up the project with dependencies and to run the test cases.

Issue-URL: https://app.clickup.com/t/85zrgkazu

Note: Also removing --verbose console tag.