fix(dapp): show new UI if zkpPrice not fetched
Without this patch, we show wrong UI
if zkpPrice not fetched successfully
in zAssets page.

This is a problem because user must be
informed if fetch has failed.

This patch solves the problem by setting
related redux state to 'failed' if
getZKPMarketPrice fetch fails, then we
show different UI to user conditionally
in zAssets page.

Issue-URL: https://app.clickup.com/t/865bmratq