fix(dapp): change unstake button color to gray while unstaking in progress
Without this patch, we have blue color unstaking
buttons in unstake tab while another unstake
action is in progress.

This is a problem because although the buttons
are disabled (not clickable), but their color
is still blue which is not good.

This patch solves the problem by adding
anotherUnstakingInProgress variable as
second condition to "locked" className.

Issue-URL: https://app.clickup.com/t/8669rbajx