feat(contracts): add Adv stake msg translator
Without this patch, There is no easy way to support new stake type
without redeploying new controller.

This is a problem because redeployment of controller
requires some updates in subgraph and FE, hence it is preferred to
create a msg translator.

This patch solves the problem by creating the
`AdvancedStakeActionMsgTranslator` contract.
It gets the stake message on new stake type, then replace the new stake
type with the old one and send the received message back to RewardMaster
but using the old type. RewardMaster then routes the message to the
old/existing controller.

Issue-URL: https://app.clickup.com/t/85zrv6qwh