feat(dapp): extend Advanced Staking to V4
This commit extends the "Advanced Staking" feature by adding support for
a new version - "Advanced Staking V4". This new version is expected to
be used for an additional 4 months.

The changes include:
-  Updating the staking term in the production environment file.
-  Adding the new staking term in the stake-terms constants.
-  Including the new staking term in the list of terms fetched by the
   getStakeTerms function.
-  Adjusting the rewards calculation function to include the new staking
   term.
-  Updating the StakeTypes type and StakeType enum to include the new
   staking term.

These changes are necessary to continue providing the Advanced Staking
feature to users, with updated terms and conditions.